backend-server/fastapi/requirements.txt
fastapi==0.115.12
uvicorn==0.33.0
sqlalchemy==1.4.15
pymysql==1.1.1
sentence-transformers==3.0.1
faiss-gpu==1.7.2
numpy==1.24.4
pandas==1.3.4
torch>=1.11.0
cupy-cuda12x>=12.0.0
httpx>=0.24.0
langchain==0.3.27
langchain-community==0.3.16
openai>=1.0.0
langchain-openai>=0.1.0
transformers>=4.35.0
accelerate>=0.25.0
